rename ThrowableHandler to ThrowableResponseFactory
